fix(hooks): move use modals to dir 'hooks'
