+++ src/pages/Typing/index.tsx
@@ -10,7 +10,7 @@ import Phonetic from 'components/Phonetic'
 import PronunciationSwitcher from './PronunciationSwitcher'
 import { isLegal } from 'utils/utils'
 import { useHotkeys } from 'react-hotkeys-hook'
-import { useModals } from 'utils/hooks'
+import { useModals } from 'hooks/useModals'
 import useSwitcherState from './hooks/useSwitcherState'
 import Switcher from './Switcher'
 import DictSwitcher from './DictSwitcher'
